refactor: 💡 contact handler is compiled into usehooks
